+++ railties/lib/rails/commands/dbconsole/dbconsole_command.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "active_support/deprecation"
+require "active_support/core_ext/string/filters"
 require "rails/command/environment_argument"
 
 module Rails
@@ -89,15 +91,15 @@ module Rails
 
     def config
       @config ||= begin
-        # We need to check whether the user passed the connection the
+        # We need to check whether the user passed the database the
         # first time around to show a consistent error message to people
         # relying on 2-level database configuration.
-        if @options["connection"] && configurations[connection].blank?
-          raise ActiveRecord::AdapterNotSpecified, "'#{connection}' connection is not configured. Available configuration: #{configurations.inspect}"
-        elsif configurations[environment].blank? && configurations[connection].blank?
+        if @options["database"] && configurations[database].blank?
+          raise ActiveRecord::AdapterNotSpecified, "'#{database}' database is not configured. Available configuration: #{configurations.inspect}"
+        elsif configurations[environment].blank? && configurations[database].blank?
           raise ActiveRecord::AdapterNotSpecified, "'#{environment}' database is not configured. Available configuration: #{configurations.inspect}"
         else
-          configurations[connection] || configurations[environment].presence
+          configurations[database] || configurations[environment].presence
         end
       end
     end
@@ -106,8 +108,8 @@ module Rails
       Rails.respond_to?(:env) ? Rails.env : Rails::Command.environment
     end
 
-    def connection
-      @options.fetch(:connection, "primary")
+    def database
+      @options.fetch(:database, "primary")
     end
 
     private
@@ -156,11 +158,21 @@ module Rails
       class_option :connection, aliases: "-c", type: :string,
         desc: "Specifies the connection to use."
 
+      class_option :database, aliases: "--db", type: :string,
+        desc: "Specifies the database to use."
+
       def perform
         extract_environment_option_from_argument
 
         # RAILS_ENV needs to be set before config/application is required.
         ENV["RAILS_ENV"] = options[:environment]
+
+        if options["connection"]
+          ActiveSupport::Deprecation.warn(<<-MSG.squish)
+            `connection` option is deprecated and will be removed in Rails 6.1. Please use `database` option instead.
+          MSG
+          options["database"] = options["connection"]
+        end
 
         require_application_and_environment!
         Rails::DBConsole.start(options)